Added Debug Constructor, Takes Stream Parameter
Added a debug constructor, this constructor takes a stream
parameter. So in the main code a Stream based object like a
HarwareSerial, SoftwareSerial maybe some other things can be passed
by pointer so debug information can be piped out of the library to
a serial device defined by the user in the main code.