Debug Output Throughout, Formatting
Put debugging output throughout the whole library. It's very
memory hungry. So it's best to comment/delete what you don't need
I'm going to take the bulk of it out in the next commit, but I
wanted to put it into this commit to show the extent of debugging
you can go to.