Reducing Memory Footprint
Reducing the memrory footprint is required because we suspect there is
memeory overflow issues with the serial print function unpredictable
behaviour. All prints of constant strings were replaced with F( and buffer
sizes were reduced. This issues will have to be dealt with better, later.

The SendCommand method was also reworked to wait the full timeout due to
the mDot sending a response with delays between parts.

TODO:
  * Pass a single full size buffer around with pointers.
  * Define all commands and responses in program memeory (PROGMEM)