Getters & Setters Working
The getters and setters now work, there is a BUG in the get and set for
the network key. It seems like the mDot returns a NULL between the O and K
causing the _sendCommand function to exit due to a NULL rather than OK.
Therefore returning (arguably) the wrong answer code.

As a workaround it looks for something other than NULL in the response.

It doesn't seem to be a stack or memeory issue, because rearranging the
stack size, and memory allocation doesn't change the way it fails AT-ALL.

Does the mDot actually (and consistantly) return a NULL between the O and
K when the AT+NK[insert anything valid here] command is called?