Set Defaults on Begin()
After some discussion it was decided it's best to set the mDot to some
known default settings on begin() to try an minimse the potential ways in
which a user make experience faults. Goal is to use this library with the
least number of main code lines as practical. a begin() then setDefaults()
is not consistant with our project goal.