_pairsToJson takes jsonLength
Because we need to know the array size of the JSON buffer, pointed to by
char* json so we don't overwrite the buffer. The sizeof() instruction
gives the size in bytes of that varible. Which for an array correctly
gives the size in bytes of the full array. But for a pointer to an array
it will give the size of that pointer. In this case 2, for 2 bytes. This
is not what we want.