Multifragment Send Based on Desired Payload Use
Multiple fragments are sent based on the desired payload use, note that
the full capacity of higher payloads isn't designed to be used by this
library at this time.

The _MAX_MDOT_COMMAND buffer size needed to be increased to to fit a
possible 126 byte payload, with an 8 byte command i.e. "AT+SEND ".