Fixed Returns Added Undesirable Delay
Fixed bug with nested returns, in general returns need to be reviewed.

Added a delay just beforee the buffer flush, if the mDot harware serial is
written to and read from, too frequently in too quick succession we get
garbage sometimes. This messes things up. This small delay fixes the
issue.